backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
dnspython==2.8.0
Flask-JWT-Extended==4.7.1
flask-cors==6.0.2
Flask==2.3.3
greenlet==3.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==8.0.33
protobuf==3.20.3
pycparser==3.0
PyJWT==2.12.1
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
gunicorn==20.1.0